\name{dmrp_covars}
\alias{dmrp_covars}
\docType{data}
\title{Paper Covariate Set}
\description{
	This data set is used in the paper Microbiome Recursive Partitioning 2019.  It contains 128 subjects and 11 cytokines.
}

\usage{data(dmrp_covars)}

\format{The format is a data frame of 128 rows by 11 columns, with the each row being a separate subject and each column being a different cytokine.}

\examples{
	data(dmrp_covars)
}

\keyword{datasets}