\name{dmrp_data}
\alias{dmrp_data}
\docType{data}
\title{Paper Taxa Data Set}
\description{
	This data set is used in the paper Microbiome Recursive Partitioning 2019.  It contains 128 subjects and 29 genus level taxa.
}

\usage{data(dmrp_data)}

\format{The format is a data frame of 128 rows by 29 columns, with the each row being a separate subject and each column being a different taxa.}

\examples{
	data(dmrp_data)
}

\keyword{datasets}