\name{HMP-internal}
\alias{HMP-internal}

\alias{loglikDM}
\alias{weirMoM}
\alias{getBC}
\alias{pioest}

\alias{gaScoring}
\alias{gaCreation}
\alias{gaPlot}

\alias{hmp.pkg.env}
\alias{rpartInit}
\alias{rpartEval}
\alias{rpartSplit}
\alias{rpartCVSingle}
\alias{rpartCV}
\alias{rpartCI}
\alias{rpartCS}

\alias{Xmcupo.statistics}
\alias{Z.statistics}
\alias{T.statistics}
\alias{Xmc.statistics}
\alias{Xsc.statistics}
\alias{Xoc.statistics}
\alias{Xdc.statistics}
\alias{Xdc.statistics.MoM}

\alias{Xmcupo.statistics.Hnull.Ha}
\alias{ZT.statistics.Hnull.Ha}
\alias{Xmc.statistics.Hnull.Ha}
\alias{Xsc.statistics.Hnull.Ha}
\alias{Xoc.statistics.Hnull.Ha}
\alias{Xdc.statistics.Hnull.Ha}


\alias{DM.Rpart.Base.Old}
\alias{rpartSplitOld}

\title{Internal Functions}
\description{These files are only called from inside other functions and are therefore not documented here.}

\keyword{internal}