\name{HMP-internal}
\alias{HMP-internal}
\alias{Xdc.statistics}
\alias{Xdc.statistics.Hnull.Ha}
\alias{Xdc.statistics.MoM}
\alias{Xmc.statistics}
\alias{Xmc.statistics.Hnull.Ha}
\alias{Xmcupo.statistics}
\alias{Xmcupo.statistics.Hnull.Ha}
\alias{Xoc.statistics}
\alias{Xoc.statistics.Hnull.Ha}
\alias{Xoc.statistics.MoM}
\alias{Xsc.statistics}
\alias{Xsc.statistics.Hnull.Ha}
\alias{ZT.statistics}
\alias{Z.statistics}
\alias{T.statistics}
\alias{ZT.statistics.Hnull.Ha}
\alias{loglikDM}
\alias{weirMoM}
\alias{LLDM}

\title{Internal Functions}
\description{These files are only called from inside other functions and are therefore not documented here.}

\author{Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon}

\keyword{internal}