\name{HMP-internal}
\alias{HMP-internal}
\alias{Xdc.statistics}
\alias{Xdc.statistics.Hnull.Ha}
\alias{Xdc.statistics.MoM}
\alias{Xmc.statistics}
\alias{Xmc.statistics.Hnull.ha}
\alias{Xmcupo.statistics}
\alias{Xmcupo.statistics.H}
\alias{Xoc.statistics}
\alias{Xoc.statistics.H}
\alias{Xoc.statistics.MoM}
\alias{Xsc.statistics}
\alias{Xsc.statistics.Hnull.Ha}
\alias{ZT.statistics}
\alias{Z.statistics}
\alias{T.statistics}
\alias{loglikDM}

\title{Internal Functions}
\description{These files are for internal use only.}

\author{Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon}

\keyword{internal}