\name{HMP-package}
\alias{HMP-package}
\alias{HMP}
\docType{package}
\title{Hypothesis Testing and Power Calculations for Comparing Metagenomic Samples from HMP}
\description{Generating data matrices following Multinomial and Dirichlet-Multinomial distributions, 
Computing the following test-statistics and their corresponding p-values, and Computing the power and size 
of the tests described above using Monte-Carlo simulations.}
\details{
\tabular{ll}{
Package: \tab HMP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-27\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon
Maintainer: Patricio S. La Rosa <plarosa@dom.wustl.edu>
}
\keyword{ package }
\keyword{ internal }